Root commit
Implements a Jenkins plugin that uses AppArmor tools on build slaves to confine builds in their workspace. This prevents builds from interacting with each other or messing with the system (like writing to jenkin's ~/.ssh/authorized_keys or other fun things)

Implements two tools for the build slaves, autoarmor-genprof which is a setuid-root helper that generates and loads AppArmor profiles for Jenkins jobs, and autoarmor-wrapper which runs a given command under an AppArmor profile.

Here we go!